+++ tomcat/src/main/java/com/techcourse/handler/DefaultHandler.java
@@ -1,15 +1,16 @@
-package org.apache.coyote.http11.handler;
+package com.techcourse.handler;
 
-import static org.apache.coyote.http11.HttpStatus.NOT_FOUND;
-import static org.apache.coyote.http11.HttpStatus.OK;
+import static org.apache.coyote.HttpStatus.NOT_FOUND;
+import static org.apache.coyote.HttpStatus.OK;
 
 import com.techcourse.exception.UncheckedServletException;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.HttpRequest;
+import org.apache.coyote.HttpRequestHandler;
+import org.apache.coyote.HttpResponse;
 
 public class DefaultHandler implements HttpRequestHandler {
 
@@ -17,7 +18,7 @@ public class DefaultHandler implements HttpRequestHandler {
 
     @Override
     public void handleGet(HttpRequest request, HttpResponse response) {
-        final URL resource = getClass().getClassLoader().getResource(STATIC_FILE_PATH_PREFIX + request.getPathInfo());
+        final URL resource = getClass().getClassLoader().getResource(STATIC_FILE_PATH_PREFIX + request.getPath());
         if (resource == null) {
             response.setStatus(NOT_FOUND);
             response.setBody("파일이 존재하지 않습니다.");
@@ -27,8 +28,8 @@ public class DefaultHandler implements HttpRequestHandler {
         final File indexFile = new File(resource.getPath());
 
         try {
-            String contentType = Files.probeContentType(indexFile.toPath());
-            String responseBody = Files.readString(indexFile.toPath());
+            final String contentType = Files.probeContentType(indexFile.toPath());
+            final String responseBody = Files.readString(indexFile.toPath());
 
             response.setContentType(contentType + ";charset=utf-8");
             response.setBody(responseBody);
