+++ tomcat/src/main/java/org/apache/coyote/HttpRequestHandler.java
@@ -1,0 +1,6 @@
+package org.apache.coyote;
+
+public interface HttpRequestHandler {
+
+    void handleGet(HttpRequest request, HttpResponse response);
+}
