--- tomcat/src/main/java/org/apache/coyote/http11/handler/HttpRequestHandler.java
@@ -1,9 +1,0 @@
-package org.apache.coyote.http11.handler;
-
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
-
-public interface HttpRequestHandler {
-
-    void handleGet(HttpRequest request, HttpResponse response);
-}
